+++ src/Atoms/Card/Card.js
@@ -1,0 +1,13 @@
+import React from 'react';
+import CardWrapperStyled from './CardStyled';
+import propTypes from 'prop-types';
+
+const CardWrapper = (props) => {
+  return <CardWrapperStyled>{props.children}</CardWrapperStyled>;
+};
+
+CardWrapper.propTypes = {
+  children: propTypes.node,
+};
+
+export default CardWrapper;
