+++ src/Atoms/Card/CardStyled.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+const CardWrapperStyled = styled.div`
+  display: flex;
+  justify-content: center;
+`;
+
+export default CardWrapperStyled;
