+++ src/Pages/LoginPage/LoginPage.js
@@ -1,32 +1,10 @@
 import React from 'react';
-import styled from 'styled-components';
 import 'font-awesome/css/font-awesome.min.css';
-import InputField from '../../Atoms/InputField';
-import Card from '../../Atoms/Card';
-import PriSecBtn from '../../Atoms/PriSecBtn';
+import InputField from '../../Atoms/InputField/InputField';
+import Card from '../../Atoms/Card/Card';
+import PriSecBtn from '../../Atoms/PriSecBtn/PriSecBtn';
 import { useState } from 'react';
-
-const StyledForm = styled.form`
-  width: 343px;
-  display: flex;
-  flex-direction: column;
-  label {
-    color: #323f4b;
-    font-size: 0.7em;
-    align-self: flex-start;
-    padding: 10px 0 3px 0;
-  }
-
-  input {
-    font-family: FontAwesome;
-    color: #7b8794;
-    font-size: 1.2em;
-    height: 48px;
-    border-radius: 8px;
-    border: 1px solid #7b8794;
-    padding-left: 10px;
-  }
-`;
+import StyledForm from './LoginPageStyled';
 
 const LoginPage = () => {
   const [login, setLogin] = useState('');
